requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-groq
langserve
openai
python-dotenv
PyMuPDF
youtube-transcript-api
chromadb
fastapi
uvicorn
sse_starlette
PyMuPDF
python-multipart
langchain-text-splitters
transformers
tiktoken